agentic_onboarding_v1/requirements.txt
streamlit>=1.36.0
pandas>=2.0.0
pyyaml>=6.0.1
sqlalchemy>=2.0.29
boto3>=1.34.0
openai>=1.40.0
email-validator>=2.2.0
